Rename the Debian package from libgul17 to gul17
[why]
There is no rule that says that Debian package names have to start with
"lib" only because they contain a library. That convention only exists
for the .so files themselves because of the way the linker looks for
files. "GUL" stands for "General Utility Library", so the "library" part
is already contained inside the name, and the "lib" prefix is redundant.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>