meson: Untangle multiple tags on one commit
[why]
While the API version increase (usually?) needs to have some change in
between two versions; which means one specific commit can not have two
different API versions, that is NOT TRUE for the git version a.k.a.
Doocs version.

It is quite possible that GUL did not change in between Doocs releases,
because GUL often does not change while Doocs gets bugfix releases. This
leaves us with multiple tags on that one commit.

For example commit 5ea254d has this tags attached to it:
DOOCSVERSION_24_6_1
DOOCSVERSION_24_6_0
DOOCSVERSION_24_6_2
DOOCSVERSION_24_3_1

We use `git describe` to get the version string. Git returns the newest
matching tag for this [1]:

    Annotated tags will always be preferred over lightweight tags, and
    tags with newer dates will always be preferred over tags with older
    dates.

The problem is of course that lightweight tags have no metadata, i.e.
no date and are all equal. Unfortunately using lightweight tags is still
prevalent here, despite the attempts to bring maintainers to use full
annotyated tags.

So if the tags above are lightweight tags one more or less random will
be returned by git describe. That is not what we need.

If they were annotated tags with dates attached to them they would
naturally be ordered by tagging date, which would be presumably be in
the order of increasing version number. But alas...

[how]
Instead of using the tag returned by git describe directly all tags that
reference that one commit are examined and sorted in reverse version
order (version sort takes care to sort '24_10_0' after '24_1_0'.
Lexically sorting would instead do not 'know' about numbers and how
people order them.
From that sorted list we take the one with the highest version number
and continue with that number like before.

If for some reason the tag sorting does not work after we already have
found at least one tag (which is very unlikely) we bail out as something
is amiss.

[note]
I do not really understand how previous Doocs-gul versions could be
correctly packaged when the Doocs version is wrong for cases where the
commit did not change between releases.

[1] man git-describe

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>