meson: Remove API version mechanics
[why]
We support building with out API version or with some abstract external
version number (Doocs version). But in fact in production the API
version is never used and it is a nice idea to support standard
behavior, but for Desy it is not used and not needed and just a
liability.

[how]
Drop the API version and all special cases that arounse through the
support.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>