ThreadPool: Add get_thread_index()
[why]
There are use cases in which it is beneficial to know on which of the N
pool threads a function gets executed.

[how]
Store a thread-local thread index when launching the threads. Return
this index when called from a worker thread. For other threads, the
thread-local index gets initialized to the maximum value of size_t;
if this value is found, get_thread_index() throws an exception.

Co-authored-by: Soeren Grunewald <soeren.grunewald@desy.de>
Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>