meson: Correct so-version
[why]
This has been forgotten; with GUL14 the version has been taken from the
git tag, and that never had any suffixes but only the number triplet.
Now we take the LIBNO file directly, and that can and will have some
suffix sometimes. Like `makeDdeb` appending the platform name, or Manuel
adding some other stuff.

Although the "usual" DOOCS libraries do have a 'complex' so-version with
suffix, GUL14 did not have it; which means one CAN upgrade a library
when just the suffix changes (makeDdeb -l 2).

This is also the behavior on MacOS, where the so-version must be
numeric. We could in principle set the darwin_versions parameter in
library(), but we keep the versions the same on all platforms and
without suffixes.

[how]
It's a bit hard to split on a non-numeric character in Meson, but with
underscorify() the number of possible characters is small enough to
handle them via strip...

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>