Correct default values for Meson project options
[why]
We use Meson's defaults for options like prefix, libdir, and includedir.
Therefore, the previously listed default values are wrong.

[how]
List "typical" values instead to give the user an idea what these
options can contain. In addition, remove a reference to the Catch2
header file that is no longer included with the library.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>