Add safe_string_view()
This adds a function that, just like safe_string() creates a std::string
from a char pointer and a length, creates a string_view instead. The
function is safe against null pointers.

Potential CHANGELOG entry:

All notable changes to this project will be documented in this file.
The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).

GUL17 releases follow a calendar-oriented versioning scheme with a version number of the
form `YY.M.minor` where `YY` stands for the last two digits of the year, `M` is the number
of the month, and `minor` is a minor version number that starts from zero for each release
in that month. For instance, the first release in March 2025 was called 25.3.0, the second
could have been called 24.3.1. This is *not* semantic versioning.

- `safe_string_view()` constructs a string_view from a char pointer and a length. It is
  safe against null pointers and non-null-terminated strings. (#27)

Initial release of GUL17. The library is based on its predecessor
[GUL14](https://gul14.info/), but requires at least C++17. Backports of C++17 standard
library features (e.g. string_view, optional) are removed, and the build system is
simplified.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>