Add a bash script to help with migration from GUL14 to GUL17
The new script looks for C++ and meson.build files in a list of
directories and performs a number of simple text replacements (like
"gul14::" to "gul17::") aimed at simplifying the migration of a C++
project from GUL14 to GUL17. The process is not perfect: Especially
where "using namespace gul14" was used, manual intervention may be
necessary. Also, we have removed the practical "make_overload_set"
helper for gul14::variant. If the script detects usage of
make_overload_set, it issues a warning.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>