Allow using join() on ranges containing const char*
[why]
So far, the join() overloads for string ranges required the element type
to provide a size() function. It should be possible to use join() also
with a vector<const char*>, however.

[how]
Detect if the element type provides a size() function with a small type
trait helper. If yes, use the previous algorithm to pre-allocate a
string of the right size. If no, simply forget about preallocation and
concatenate the strings with automatic string growth.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>