Re-import OverloadSet and make_overload_set() from GUL14
[why]
Both the OverloadSet class template and the make_overload_set() function
template were included in GUL14, but removed along with the variant.h
header in the transition to GUL17. However, at least the class template
is still useful in conjunction with std::visit().

[how]
Import both the class and the function template. Rewrite OverloadSet in
C++17 style with a deduction guide so it can be constructed directly
from an arbitrary list of lambdas. Deprecate the make_overload_set()
function because it no longer offers any benefit over instantiating an
OverloadSet directly. Import the original tests for make_overload_set().

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>