Simplify libgul14 replacement, fix GUL17 dependency name
[why]
We can simply replace all occurrences of 'libgul14' inside the
meson.build files by 'gul17'. This is correct for both wrap dependencies
and pkgconfig oned. The main dependency exported by GUL17's
meson.build is still called libgul_dep, not gul17_dep, so it is not
necessary to write a special case for it.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>