GUL14 migration: Preserve gul14/catch.h includes
[why]
GUL17 does not bundle a header-only Catch2 include file anymore, so some
manual work is required to migrate a project to Catch2v3. This is out of
the scope of the GUL14 migration script. The best way of dealing with
this is, arguably, to keep including the header that is bundled with
GUL14.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>